refactor Datasets class: move all to utils; fix tests and examples; fix Network
