+++ examples/dnn_examples/CMakeLists.txt
@@ -4,7 +4,7 @@ project(dnn_example)
 find_package(BLAS)
 
 add_executable(dnn_example dnn_example.cpp)
-add_executable(dnn_example_classification_fc dnn_example_classification.cpp)
+#add_executable(dnn_example_classification_fc dnn_example_classification.cpp)
 add_executable(dnn_example_pgm dnn_example_pgm.cpp)
 
 target_link_libraries(dnn_example_pgm ${BLAS_LIBRARIES})
